pyproject.toml
[tool.poetry]
name = "ibm-service-validator"
version = "0.0.3"
description = "Extension of endpoint validator, Schemathesis."
packages = [
    { include = "src" }
]
keywords = ["pytest", "schemathesis", "openapi", "swagger", "testing", "ibm"]
classifiers = [
    "Environment :: Console",
    "Framework :: Schemathesis",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development :: Testing",
]
authors = ["Barrett Schonefeld <barrett.schonefeld@ibm.com>"]
readme = "README.md"
repository = "https://github.ibm.com/CloudEngineering/ibm-service-validator"

[tool.poetry.dependencies]
click = "^7.0"
ibm-cloud-sdk-core = "^2.0"
python = "^3.6"
pyyaml = "^5.3"
schemathesis = "^1.7"

[tool.poetry.dev-dependencies]
coverage = "^4.5"
pytest = ">4.6.4"
pytest-mock = "^1.11.0"
typing_extensions = "^3.7"

[tool.black]
line-length = 90
target_version = ["py37"]

[tool.isort]
# config compatible with Black
line_length = 90
multi_line_output = 3
default_section = "THIRDPARTY"
include_trailing_comma = true
known_first_party = "ibm_service_validator"
known_third_party = ["_pytest", "click", "flask", "hypothesis", "ibm_cloud_sdk_core", "ibm_service_validator", "pytest", "requests", "schemathesis", "setuptools", "yaml"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
